pyproject.toml
[tool.poetry]
name = "election"
version = "0.1.0"
description = ""
authors = ["Daniel Bibin <danielbibin9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
streamlit-option-menu = "^0.4.0"
pillow = "^11.3.0"
mysql-connector-python = "^9.3.0"
streamlit = "^1.47.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
